projet_ovirt/requirement.txt
appdirs==1.4.3
argcomplete==1.9.4
attrs==19.1.0
backcall==0.1.0
cached-property==1.5.1
certifi==2018.11.29
chardet==3.0.4
config==0.4.0
decode==0.5.1
decorator==4.3.2
defusedxml==0.5.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.13.3
lxml==4.3.2
ovirt-engine-sdk-python==4.3.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
policies==0.4.2
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py-zabbix==1.1.5
pybase64==0.5.0
pycurl==7.43.0.2
Pygments==2.3.1
pylibcurl==0.7.12
pyparsing==2.3.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
zeep==3.2.0
